[pdf] Change structure identifiers to match specification
Many of the types and identifiers currently used to create the logical
structure of a PDF document are confusingly different from the terms
used in the specification. In particular "node" and "tag" are used for
concepts quite different from what these terms are used for in the code.
Change these identifiers to have names closer to those used by the
specification. Only private identifiers are changed, public identifiers
should also be updated in a future change.

Identifier changes:
SkPDFTagTree to SkPDFStructTree (spec "StructTreeRoot")
tagTree to structTree (spec "structure tree")
SkPDFTagNode to SkPDFStructElem (spec "StructElem")
tag to structElem (spec "structure element")
nodeId to elemId (spec "element identifier")
markId to mcid (spec "marked-content identifier" or "MCID")
AnnotationInfo to ContentItemInfo (spec "content item")
markUsed to setUsed (to avoid confusion with "mark")
SkPDFTagTree::PrepareTagTreeToEmit to SkPDFStructElem::emitStructElem
nodeMap to structElemForElemId
marksPerPage to structElemForMcidForPage
parentTreeAnnotationNodeIds to structElemForContentItem

SkPDFTagTree::fIdTreeEntries is now the local idTree in
SkPDFStructTree::emitStructTreeRoot, as it was not used outside of this
method.

Comments updated to reflect the new identifiers and document the methods
more in terms of the specification.

SkPDFTagTree::addNodeAnnotation is folded into
SkPDFStructTree::createStructParentKeyForElemId. It was discovered to be
quite difficult to write proper documentation for
SkPDFTagTree::createStructParentKeyForNodeId and
SkPDFTagTree::addNodeAnnotation since it turns out these were two halves
of a single operation. The former was used to associate a content item
with a structure element and the later was used to associate the
structure element with the content item. Instead, keep things consistent
by doing both in SkPDFStructTree::createStructParentKeyForElemId. The
one user is updated trivially by reserving a content item reference
before emitting the content item. However, this change revealed that
createStructParentKeyForNodeId has not actually done anything for some
time, since it first checks to ensure there is an active page (since the
parent key goes with the page) but SkPDFDocument::endPage closed the
page before processing the annotations on the page. So this change must
also change SkPDFDocument::endPage to close the page later.

Change-Id: I881aa13370efc68565c9c5bbba884964c98f1cc8
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/911656
Commit-Queue: Ben Wagner <bungeman@google.com>
Reviewed-by: Florin Malita <fmalita@google.com>